Javadoc updates

+++ MoC-Domain/src/main/java/domain/HintComparator.java
@@ -2,10 +2,15 @@ package domain;
 
 import java.util.Comparator;
 
+/**
+ * Compares two hints dependant on when they are scheduled to be released.
+ *
+ * @author TeamKoekje
+ */
 public class HintComparator implements Comparator<Hint> {
 
     @Override
     public int compare(Hint t, Hint t1) {
-        return Integer.compare(t.getTime(), t1.getTime());
+        return Long.compare(t.getTime(), t1.getTime());
     }
 }
